backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
python-multipart==0.0.20
pydantic==2.12.4
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
opencv-python==4.12.0.88
pillow==12.0.0
requests==2.32.5
geopy==2.4.1
groq>=0.12.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.20

